separate logic of PosterItem component
