+++ src/components/Catalog/Posters/PosterItem/models.ts
@@ -1,0 +1,18 @@
+export interface PosterItemProps {
+  setMovieDetails: {
+    title: string;
+    tagline: string;
+    vote_average: number;
+    vote_count: number;
+    release_date: string;
+    poster_path: string;
+    overview: string;
+    budget: number;
+    revenue: number;
+    runtime: number;
+    genres: string[];
+    id: number;
+  };
+  poster: any;
+  genre: any;
+}
