+++ src/components/Catalog/Posters/index.tsx
@@ -1,57 +1,29 @@
-import React, { FC, useCallback, useEffect } from 'react';
+import React, { FC, useCallback, memo } from 'react';
 import shortid from 'shortid';
 
 import { PostersProps } from './models';
 import {
   StyledPostersWrapper,
   StyledPostersList,
-  StyledPostersItem,
-  StyledPostersWrapTitle,
-  StyledPostersTitle,
-  StyledPostersTitleYear,
-  StyledPostersImg,
-  StyledPostersGenre,
-  StyledPostersLink,
   StyledNumberMovies,
   StyledPostersError,
 } from './style';
 
 import usePostersFetch from 'hooks/usePostersFetch';
-import useMoreDetailsMovie from 'hooks/useMoreDetailsMovie';
-
-import defaultImgMovie from '../../../assets/images/fallback_movie.png';
 
 import { API_PAGE } from '@constants';
 
 import Button from 'components/Button';
 import { Spinner } from 'components/Spinner';
+import PosterItem from './PosterItem';
 
 const Posters: FC<PostersProps> = ({ setMovieDetails }: any) => {
   const { movies, error, loading, fetchMovies }: any = usePostersFetch();
-  const {
-    movieDetails,
-    loadingMovieDetails,
-    errorMovieDetails,
-    fetchMovieDetails,
-  }: any = useMoreDetailsMovie();
 
   const loadMorePosters = useCallback(() => {
     const loadMorePostersEndpoint = `${API_PAGE}${movies.currentPage + 1}`;
 
     fetchMovies(loadMorePostersEndpoint);
-  }, []);
-
-  const handleMoreDetails = useCallback((id) => {
-    fetchMovieDetails(id);
-  }, []);
-
-  useEffect(() => {
-    setMovieDetails(movieDetails);
-  }, [movieDetails]);
-
-  const addDefaultSrc = useCallback(({ target }) => {
-    target.src = defaultImgMovie;
-    target.alt = 'image not found';
   }, []);
 
   const posters = movies.movies.map((poster) => {
@@ -60,22 +32,12 @@ const Posters: FC<PostersProps> = ({ setMovieDetails }: any) => {
     ));
 
     return (
-      <StyledPostersItem key={shortid.generate()}>
-        <StyledPostersLink onClick={() => handleMoreDetails(poster.id)}>
-          <StyledPostersImg
-            src={poster.poster_path}
-            alt={poster.title}
-            onError={addDefaultSrc}
-          />
-          <StyledPostersWrapTitle>
-            <StyledPostersTitle>{poster.title}</StyledPostersTitle>
-            <StyledPostersTitleYear>
-              {poster.release_date}
-            </StyledPostersTitleYear>
-          </StyledPostersWrapTitle>
-          <StyledPostersGenre>{genre}</StyledPostersGenre>
-        </StyledPostersLink>
-      </StyledPostersItem>
+      <PosterItem
+        key={shortid.generate()}
+        setMovieDetails={setMovieDetails}
+        poster={poster}
+        genre={genre}
+      />
     );
   });
 
@@ -100,4 +62,4 @@ const Posters: FC<PostersProps> = ({ setMovieDetails }: any) => {
   );
 };
 
-export default Posters;
+export default memo(Posters);
